fixing size_t and int problems

+++ computerservice.cpp
@@ -8,7 +8,7 @@ ComputerService::ComputerService()
 
 }
 
-Computer ComputerService::getComputer(int index)
+Computer ComputerService::getComputer(size_t index)
 {
     return _computers[index];
 }
@@ -24,7 +24,7 @@ void ComputerService::retrieveComputers()
     _computers = data.getComputers();
 }
 
-int ComputerService::getSize() const
+size_t ComputerService::getSize() const
 {
     return _computers.size();
 }
